backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
python-dotenv==1.0.0
facebook-business==19.0.3
pandas>=2.2.0
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
python-dateutil==2.8.2
loguru==0.7.2